chore: move last distributions to all Panel
