--- src/components/common/lastDistribution/styles.module.scss
@@ -1,17 +1,0 @@
-@import 'assets/styles/_mixins.scss';
-
-.container{
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  padding: 10px;
-  font-size: 12px;
-  color: gray;
-  min-width: 130px;
-
-  .loader{
-    width: 20px;
-    background: none;;
-  }
-}
