+++ src/components/layout/PanelChange/styles.module.scss
@@ -146,6 +146,13 @@
   color: var(--color-grey);
 }
 
+.distributed_time {
+  font-style: normal;
+  font-size: 12px;
+  line-height: 100%;
+  color: var(--color-grey);
+}
+
 .coin {
   width: 100%;
   flex: 1;
